fix: address bugbot ticket and checkout findings
Avoid duplicate ticket lifecycle syncs, repair attendee question storage independently, and centralize donation classification so checkout totals exclude all donation order items consistently.
